requirements.txt
discord.py
python-dotenv
Flask
motor
certifi
